services/loadgen/requirements.txt
httpx==0.27.2
websockets==12.0
numpy==1.26.4
psutil==6.0.0
PyYAML==6.0.2
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4